requirements.txt
fastapi[standard]
pydantic
pydantic-settings
bcrypt
python-jose[cryptography]
python-multipart
sqlalchemy
aiosqlite
